+++ DTCD-SDK/DatasourceClasses/index.js
@@ -1,0 +1,9 @@
+import { BaseDataset } from './BaseDataset';
+import { BaseDatasourceContent } from './BaseDatasourceContent';
+import { BaseExternalDatasource } from './BaseExternalDatasource';
+
+export {
+  BaseDataset,
+  BaseDatasourceContent,
+  BaseExternalDatasource,
+};